Add config option to enable natsemi net driver hardware bug workaround.
"some" systems with "some" cables see a large amount of errors,
due to a hardware bug.  This bug is (apparently) not probe-able;
however it only appears on rare reference boards and the like,
so we simply add a config option and default the option to OFF.

Further detail:
When CONFIG_NATSEMI_CABLE_MAGIC option is enabled, PMDCSR_VAL
register value becomes 0x1898, a value provided by a NatSemi
app note.  This enables a workaround for a hardware bug
which is (apparently) not probe-able.  Luckily the hardware bug
is (apparently) not common either, so we default to disabling
this workaround.

Contributor: Tim Hockin