[PATCH] Optimized UP preempt fix
I previously sent a patch by Mikael Pettersson to fix the UP+preempt
problem.  It seems from your BK repository you have not yet merged it;
if so, this patch takes a different approach which is optimal, removing
the unneeded conditional altogether in the UP case.  I have verified UP
and SMP are now correct.  Patch is against 2.5.4, please apply.

	Robert Love