[PATCH] Re: IDE cleanup for 2.5.4-pre3
The end_request() function familiy (not the global one, but the IDE
specific ones), did bear a permuted parameter ordering.  After fixing
this it turned out that at all places the huk parameter wasn't the
hwgroup, but just the drive in question itself.  I have changed this to
be more sane, which allowed to remove many unneccessary code
duplication, or rather obfuscation, in between the __ide_end_request()
and ide_end_request() functions.  This simplification is actually the
"spreading" part of the game.