[PATCH] New wireless driver API part 2
	Quick summary : this patch build on the first part to offer
two important new features :
		o Wireless Events
		o Wireless Cell Scanning
	Wireless Events are events generated by device, driver or the
wireless subsystem. It allows for example a device to notify user
space when it register to a new cell (roaming) or loose contact with
the current Access Points. Currently, the other defined events include
some configuration changes and packet drop due to excessive retries,
more may come in the future. All those events are useful for MobileIP,
V-Handoff and Ad-Hoc routing.
	Wireless Cell Scanning is a generic API to allow device/drive
to report Wireless Cells discovered (including ESSID, frequency and
QoS). This is similar to what is available in WindowsXP (except that
it's compliant to Wireless Extensions).

	This patch has been submitted for review on this list a couple
of time in January, has been on my web page since and used intensively
by other people. It was rediffed to 2.5.6. Driver patches have been
submitted to maintainers.