[PATCH] 2.5.6 correct NFS client handling of EJUKEBOX error...
  The following patch resyncs 2.5.6 with the 2.4.x series w.r.t. the
handling of the EJUKEBOX error. The latter is an NFS-specific error
that is returned by servers that support hierarchical storage: it
notifies the client that the request cannot be completed in a timely
fashion (Imagine for instance a situation where you have a cdrom
jukebox system, and the user has just requested a file on another cd).

Under these circumstances, the RFC specifies that the request should
be retried after suitable timeout during which the server will attempt
to perform whatever action is required to make the file available
again.