e100 net driver update:
1) This patch provides fix for "wake on arp" and "wake on unicast"
functionality when card is suspended by power management. When e100_suspend
was called for the device that is in netif_running state the load filter
command was executed in the asynchronic mode and the order of actions
required to put device into wake up enabled mode was broken.

The fix enables to execute WOL configure and load filter commands in the
synchronic mode despite of fact that device is in netif_running state. The
exec_non_cu_command uses the driver_isolated flag to identify this
situation. 

2) add EXPORT_NO_SYMBOLS (yay Intel, you have come so far... :))

3) bump version to 2.0.25-pre1