[PATCH] kjournald exits with nonzero preempt_count
The preempt_count debug check that went into 2.5.8-pre1 already caught a
simple case in kjournald.  Specifically, kjournald does not drop the BKL
when it exits as it knows schedule will do so for it.

For the sake of clarity and exiting with a preempt_count of zero, the
attached patch explicitly calls unlock_kernel when kjournald is exiting.