[PATCH] Fix console initialization deadlock for x86-64
When vt_init happens to run before the keyboard is initialized (the order is random
because they're both different initcalls) then set_leds raises an not yet enabled
tasklet. This causes an endless loop on the first schedule() call because the tasklet
handling cannot handle raised but disabled tasklets.

This patch just does not do set_leds in virtual terminal initialization to avoid that.
It is done later anyways.