[PATCH] PATCH 2.5.10 usbnet, check_connect updates
USB usbnet driver update

The patch cleans up the handling of connection checking to
match the truth that it's effectively tristate (connected, not,
and can't know) vs binary ("connected" and "can't know"
as the same value, since it was OK to ifup).  Since ethtool
supports the tristate semantics, it matters now.