[PATCH] Re: ohci-hcd and that NCR implementation ...
The attached patch makes the one SiS controller behave
like the other... the key fix being wmb() to make sure that
no writes are still in CPU caches before telling the HC it's
OK to fetch the data.  (And a minor tweak to access PCI
less often through the hub driver.)

But the OPTi problem seems to be different.  Could you
try this patch to see if it affects the problem your seeing
with that NCR implementation?  And in any case, please
integrate it to Linus' latest.  (Against 2.5.10, might be
some fuzz on the pci tweak due to that warning fix.)

Those missing wmb() statements have likely been the root
cause of various usb-ohci flakes reported over time, too.
The drivers are still almost identical in those areas.  (Though
something made ohci-hcd consistently lose that cache flush
race; maybe it's a bit faster.)