[PATCH] - kNFSd in 2.5.15 - export_operations support for isofs
ISOFS to use export_operations to declare exportability

Exporting ISO's has never been very reliable
as lookup("..") doesn't work (even though there is
some code that tries to pretend that it does).

This means that if an inode falls out of cache, any
filehandle on it will become stale.

This makes that level of support explicit with an empty
export_operations structure.  get_parent could be written....
but not today.